Replace pitches-ises.ly with real music
This actually does a much better job at demonstrating what the paragraph
below said: the fis doesn’t get a sharp because it’s part of the key
signature, and the f gets a natural. Which is good, because the
paragraph has to be significantly shortened to make room for the extra
code (key and time signature).

This also means that we now use some duration notation before
introducing it, but I’ll move the duration slide ahead in the next
commit, it doesn’t use pitch anyways.